Reset path on shell load
Why:
Was having issues with loading the correct ruby env (gem env)
related to the path_helper executable. There was also some unused
code originally needed for managing local config vs. tb/dotfiles standard

This PR:
Adds a check if a login shell (not in tmux) and if there is a path_helper
exe, if so, reset the path manaully so that we don't have duplications in
the path.