Update tmux conf to have status bar top
Why:
We want to have the tmux status bar at the top of the tmux session
window. When switching over to the alacritty terminal there was a screen
repainting issue with tmux when the status bar was at the bottom of the
screen, putting at the top fixes this. Additionally there is some visual
separation from the vim status bar and the tmux status bar when they are
being used at the same time.