Add galias alias
Why:

We would like to be able to show the list of available git aliases with
a single command in the terminal. While we currently have the `git
alias` command, we would like to pretty print these aliases for quality
of life purposes.

This commit:

Adds a new alias, 'galias', which pretty prints the current git aliases.