Add ll zsh function
Why:

We would like to have a way to show all files, executables, and
directories listed in director, executable, file and then alphabetical
order.

This commit:

Adds a `ll` function which acts like ls -oa, but ordered with
directories and executables on top.

Note this is only required on non-GNU shells per this [stack exchange
post](https://unix.stackexchange.com/questions/22815/how-to-list-files-and-directories-with-directories-first)