Use init.lua for nvim
Why:

Recent versions of nvim have fuller support for configuration written in
lua. Lua is a better language for embedded scripting and the vim
community has been moving from vimscript to lua. We'd like to update
these dot files to stay up to date with the industry.

This commit:

Removes init.vim and adds init.lua.

Most of the configuration is the same, but we did update a few things:
- updated some key mappings to match Heiker Curiel's setup
- removed settings which were duplicating defaults
- replaced vim-plug to lazy.nvim
- removed unused plugins
- replaced ack with ripgrep

https://vonheikemen.github.io/devlog/tools/build-your-first-lua-config-for-neovim/
https://vonheikemen.github.io/devlog/tools/neovim-plugins-to-get-started/