Use consistent comment header conventions
Why:

We'd like to update our zshrc file to use the same header comment
conventions as we use in init.lua as it improves readability when moving
between config files when we are consistent.

This commit:

Updates header comments in zshrc to use the following format:

```
```

We also add a commented out simple version of the shell prompt for use
in making screen recordings of terminal demonstrations.