Add setup for linux
Why:

We'd like to be able to quickly setup dotfiles for linux machines.

This commit:

Adds a markdown file documenting the setup for linux machines.

Updates the setup_dotfiles script to work for both linux and mac
machines.

Adds configuration for the i3 window manager.

We additionally remove no longer tools like ack, which was replaced with
ripgrep, and vimrc, which was replaced with vim/init.lua