Add volume and brightness controls for i3
Why:

Given we are using i3 for our window manager we need to provide a way to
addjust the volume and screen brightness as those are not provided out
of the box with i3.

This commit:

Adds a few bash scripts with corresponding shell aliases for setting the
volume of the speakers and brightness of the screen.

Note that a future commit could update the i3 config to bind the
keyboard keys to these scripts.